requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-core>=0.1.6
langchain-text-splitters>=0.0.1
chromadb>=0.4.0
sentence-transformers>=2.2.0
transformers>=4.35.0
torch>=2.0.0
python-docx>=0.8.11
python-pptx>=0.6.21
PyPDF2>=3.0.1
pydantic
FastMCP
